requirements.txt
flask==2.3.0
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
pyjwt==2.8.0
python-dotenv==1.0.0
werkzeug==2.3.0
torch==2.6.0
torchvision==0.21.0
pillow==10.0.0